Deprecate GenerateOptions
