+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/generator/ArrayContainerPropertyGenerator.java
@@ -40,7 +40,7 @@ public final class ArrayContainerPropertyGenerator implements ContainerPropertyG
 
 		ArbitraryContainerInfo containerInfo = context.getContainerInfo();
 		if (containerInfo == null) {
-			containerInfo = context.getGenerateOptions()
+			containerInfo = context.getFixtureMonkeyOptions()
 				.getArbitraryContainerInfoGenerator(property)
 				.generate(context);
 		}
