+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/generator/DefaultSingleContainerPropertyGenerator.java
@@ -57,7 +57,7 @@ public final class DefaultSingleContainerPropertyGenerator implements ContainerP
 
 		ArbitraryContainerInfo containerInfo = context.getContainerInfo();
 		if (containerInfo == null) {
-			containerInfo = context.getGenerateOptions()
+			containerInfo = context.getFixtureMonkeyOptions()
 				.getArbitraryContainerInfoGenerator(property)
 				.generate(context);
 		}
