+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/generator/MapContainerPropertyGenerator.java
@@ -55,7 +55,7 @@ public final class MapContainerPropertyGenerator implements ContainerPropertyGen
 
 		ArbitraryContainerInfo containerInfo = context.getContainerInfo();
 		if (containerInfo == null) {
-			containerInfo = context.getGenerateOptions()
+			containerInfo = context.getFixtureMonkeyOptions()
 				.getArbitraryContainerInfoGenerator(property)
 				.generate(context);
