+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/generator/StreamContainerPropertyGenerator.java
@@ -51,7 +51,7 @@ public final class StreamContainerPropertyGenerator implements ContainerProperty
 
 		ArbitraryContainerInfo containerInfo = context.getContainerInfo();
 		if (containerInfo == null) {
-			containerInfo = context.getGenerateOptions()
+			containerInfo = context.getFixtureMonkeyOptions()
 				.getArbitraryContainerInfoGenerator(property)
 				.generate(context);
 		}
