+++ fixture-monkey/src/main/java/com/navercorp/fixturemonkey/tree/ObjectTree.java
@@ -41,7 +41,7 @@ import com.navercorp.fixturemonkey.api.generator.ArbitraryProperty;
 import com.navercorp.fixturemonkey.api.generator.CombinableArbitrary;
 import com.navercorp.fixturemonkey.api.generator.FixedCombinableArbitrary;
 import com.navercorp.fixturemonkey.api.matcher.MatcherOperator;
-import com.navercorp.fixturemonkey.api.option.GenerateOptions;
+import com.navercorp.fixturemonkey.api.option.FixtureMonkeyOptions;
 import com.navercorp.fixturemonkey.api.property.Property;
 import com.navercorp.fixturemonkey.api.property.RootProperty;
 import com.navercorp.fixturemonkey.customizer.NodeManipulator;
@@ -50,7 +50,7 @@ import com.navercorp.fixturemonkey.customizer.NodeManipulator;
 public final class ObjectTree {
 	private final RootProperty rootProperty;
 	private final ObjectNode rootNode;
-	private final GenerateOptions generateOptions;
+	private final FixtureMonkeyOptions fixtureMonkeyOptions;
 	private final ObjectTreeMetadata metadata;
 	private final MonkeyContext monkeyContext;
 	@SuppressWarnings("rawtypes")
@@ -60,16 +60,16 @@ public final class ObjectTree {
 	public ObjectTree(
 		RootProperty rootProperty,
 		ObjectNode rootNode,
-		GenerateOptions generateOptions,
+		FixtureMonkeyOptions fixtureMonkeyOptions,
 		MonkeyContext monkeyContext,
 		List<MatcherOperator<? extends FixtureCustomizer>> builderCustomizer
 	) {
 		this.rootProperty = rootProperty;
 		this.rootNode = rootNode;
-		this.generateOptions = generateOptions;
+		this.fixtureMonkeyOptions = fixtureMonkeyOptions;
 		this.monkeyContext = monkeyContext;
 		List<MatcherOperator<? extends FixtureCustomizer>> concat = new ArrayList<>();
-		concat.addAll(generateOptions.getArbitraryCustomizers());
+		concat.addAll(fixtureMonkeyOptions.getArbitraryCustomizers());
 		concat.addAll(builderCustomizer);
 		this.customizers = concat;
 		MetadataCollector metadataCollector = new MetadataCollector(rootNode);
@@ -154,7 +154,7 @@ public final class ObjectTree {
 			if (node.isNotManipulated() && notCustomized && cached != null) {
 				generated = cached;
 			} else {
-				generated = this.generateOptions.getArbitraryGenerator(node.getResolvedProperty())
+				generated = this.fixtureMonkeyOptions.getArbitraryGenerator(node.getResolvedProperty())
 					.generate(childArbitraryGeneratorContext);
 
 				if (node.isNotManipulated() && notCustomized) {
